% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip_fucntion.R
\name{mip}
\alias{mip}
\title{Maximal Intensity Period}
\usage{
mip(Metric, dur)
}
\arguments{
\item{a}{Colum "Metric" of the data frame that wants to be analysed}

\item{b}{Duration of the maximum intensity period in seconds}
}
\value{
Numeric value
}
\description{
This function is used to extract the maximal intensity output of a variable for a specific duration
}
\examples{
mip(data$MetabolicPower,60)



}
