% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data for analysis}
\format{
An object of class \code{data.frame} with 27635 rows and 4 columns.
}
\usage{
sample_data
}
\description{
This is a dataframe with example 10Hz GPS data, with calculated columns of Speed in meters per minute, Metabolic Power and Acceleration density. It can be used to run the examples of the functions
}
\examples{
data <- sample_data

}
\keyword{datasets}
