% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Submip_function.R
\name{submip}
\alias{submip}
\title{Submaximal Intensity Periods}
\usage{
submip(MIP, dur, Metric, threshold)
}
\arguments{
\item{MIP}{Reference value of maximum intensity period, either calculated with mip function or previous/maximal player value}

\item{dur}{Duration of the maximum intensity period in seconds}

\item{Metric}{Colum "Metric" of the data frame that wants to be analysed}

\item{threshold}{Threshold for the detection of the submaximal intensity period}
}
\value{
A dataframe with 3 columns, id, duration and intensity of each submaximal intensity period
}
\description{
This function is used to extract the submaximal intensity periods that are above a specific MIP threshold
}
\examples{
submip(19.6,60,data$MetabolicPower,90)
}
