% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submip_plot_function.R
\name{submip_plot}
\alias{submip_plot}
\title{SubMaximal Intensity Periods Plot}
\usage{
submip_plot(MIP, dur, Metric, threshold)
}
\arguments{
\item{MIP}{Reference value of maximum intensity period}

\item{dur}{Duration of the maximum intensity period in seconds}

\item{Metric}{Colum "Metric" of the data frame that wants to be analysed}

\item{threshold}{Threshold for the detection of the submaximal intensity period}
}
\value{
plot
}
\description{
This function is used to plot the submaximal intensity periods that are above a specific MIP threshold, locate them in the time series and visualize its duration and intensity
}
\examples{
submip_plot(19.6,60,data$MetabolicPower,90)
}
